Feature/player UI

+++ Assets/ScriptableObject/EnemyScriptableObject.cs
@@ -15,6 +15,6 @@ public class EnemyScriptableObject : ScriptableObject
     public float MaxHealth { get => maxHealth; private set => maxHealth = value;}
 
     [SerializeField]
-    float damage;
-    public float Damage{get => damage; private set => damage = value;}
+    int damage;
+    public int Damage{get => damage; private set => damage = value;}
 }
